JSON types: remove "_path must be first column" check
This commit removes the validation check that ensured that _path was
present as the first field of descriptors. This is a reflection of our
zeek-onlyness of when this feature was first introduced. Now with the
impending arrival of Suricata and other data types, this hardcoded
check no longer belongs.

If we find ourselves needing such a validation, we can re-introduce
something more general once we have a few more ingest types under our
belt.