backward compat for reading the alpha version of zng files
This commit adds backward compatibility for reading alpha-version
zng files.  Right now, the tip of the tree is still the alpha version
but we will soon introduce changes so we are putting this backward
compat in now, in front of coming changes.

This commit introduces a copy of package zng and other bits we need
under the zq/alpha tree.  This can be removed down the road when we
fully deprecate the alpha zng format.

Since the tzng format hasn't changed except for "byte" to "uint8",
which isn't currently easy to get into a zng file without hand editing,
we use a trip from old zng to tzng to new zng to do the backward compat.