fix field model and dot expression syntax
This commit fixes problems with the design patterns underlying
fields, dotted accesses, and the data structures in packages
ast and expr supporting this.  Dots can now be in field names
and the field representation is no longer a string, but rather
a slice of strings (as defined in the new field package in zq/field).

For example, you can now say .["@."] to refer to the field named "@."
in the root record refered to with '.'.  'foo' is shorthand for '.foo'
and 'foo["bar"]' is, the same as 'foo.bar', but 'foo[bar]' means to
look for a field in record '.foo' whose name is the value of '.bar'.

nYou can also say things like 'count() by ." or "count() by typeof(version)"

Reducer and assignment expressions within the ast have been simplified
and follow more conventional compiler design patterns.  Also,
field accesses now explicitly start from the root record and there
is now an ast.RootRecord that represents this.

This refactoring allowed several other improvements:
- You can sort by expressions, e.g., "sort String.toLower(s)"
- Put expressons are flexible, e.g., "put typeof(.),String.toLower(s)"
- Cut expressions are more flexible, e.g., "cut x=y+1,typeof(s)"

Finally, the PEG grammar was simplified and made more orthogonal
so it will be easier to extend and modify moving forward.