backward compat for reading the alpha version of zng files (#1386)
This commit adds backward compatibility for reading alpha-version
zng files.  Right now, the tip of the tree is still the alpha version
but we will soon introduce changes so we are putting this backward
compat in now, in front of coming changes.

This commit uses v0.22.0 of package zng using go module awesomeness 
via a "replace" directive in go.mod to provide the backward compat.

Since the tzng format hasn't changed except for "byte" to "uint8",
which isn't currently easy to get into a zng file without hand editing,
we use a trip from old zng to tzng to new zng to do the backward compat.