fix bug in spill file name allocation (#1391)
The recent refactor of the spill code into its own module introduced
a bug where the file names could collide because they were named
after their position in the spill table, which grows and shrinks.
This caused distinct spills to have the same file name.  The fix is
to track the total number of spills and name the file after this.