refactor package expr (#1375)
This commit refactors the expr package to be more extensible and
performant.  The zngnative package has been eliminated.  It had
good intentions but every expression resulted in the allocation
of an empty interface to hold a go native value resulting in lots
of traffic to the GC.  This implementation keeps scratch buffers
around and reuses them and aside from strings and IPs, does not
allocate any memory to evaluate expressions. 

We also changed the field implementation and cleaned up the AST.
There was overlap between the old FieldCall stuff and the generic
expression code, which is now eliminated.  This also makes the
grammar a bit easier to reason about as we improve and extend it.

There is now an interface type, expr.Evaluator, to take over for
all the function closures we previously had.  This makes it
easier to cookie-cutter and add new expression functionality.
Also, the field expressions are now just an instance of
expr.Evaluator so everything is now clean and uniform.

While we were changing the coercion logic around and getting
rid of zngnative, we laid out the ZNG type codes a bit differently
and updated the ZNG spec.  Since we were making incompatible changes,
we changed the "byte" type to "uint8" and added "int8" so all
the integer types are now orthogonal.  Also, the integers are
all encoded using the same varint logic now, which greatly
simplifies value encoding and decoding in the coercion engine.

Most of the tests that were updated simply accommodate the change
to the ZNG binary format.