update zjson spec and implementation (#1299)
This commit updates the zjson spec to conform with feedback from
our experience and use.  The main change is to format types
completely using structured JSON instead of a mix of structure
for record types and strings for array/set/union.  Likewise,
we fixed the aliases to be coded as structured JSON instead
of type strings.