add ast command (#1416)
This commit adds a new command called "ast", which prints parsed zql as its
underlying JSON object. This is useful for test and dev and debugging zql queries.
As we add more sophisticated ast transformations as various optimizations are
implemented, this command could prove very useful for test, e.g., writing ztest-style
tests for the various ast transformations. It's also useful for debugging other parts
of the system to dig into how the ast is being transformed for distributed processing.
In this way, it's a bit like sql "EXPLAIN".