add where clause to reducers (#1490)
This commit adds a "where" clause to reducers allowing you to
apply filtering to individual aggregations in a multi-column
group-by.  e.g.,

  lg=sum(foo) where x>100, sm=sum(foo) where x<100 by bar

This was also the right time to simplify the reducer syntax in the
peg grammar so that there is no longer an enumerated list of the available
aggregation functions.  Instead, the compiler pass makes sure the
the named reducer exists making it easier from now on to add reducers
without having to modify the grammar.