fix bug with groupby ast
This commit fixes a bug in the groupby ast where the limit arg wasn't
being passed through without a keys argument (arguably the limit isn't
needed without keys, but that will no longer be the case for large-valued
reducers using union or collect).

We also simplify the groupby ast assembly logic in the peg rule by
following the pattern that any optional args with leading space
have space in the child rule not the parent.

Also, we added several omitempty tags to make the go ast json more
in line with the javascript json.