Add api.SearchRequest.ZQL as a .Proc alternative
A zqd client must provide an AST in api.SearchRequest.Proc, requiring
the client to implement a ZQL parser if it is to allow searches with
arbitrary ZQL.  Add api.SearchRequest.ZQL to lift this requirement.
Either the Proc field or the ZQL field may be provided in requests to
zqd, but providing both is an error.