change zng type error from native type to alias
This commit removes the "error" type from the zng spec and updates
the zq implementation to use an alias type for "error" (aliased
to string) instead of the native type.

This requires passing the type context into the expression compiler.
Right now, we've changed the type signature of expr.Function type
to take the type context, but in a future commit, we will change
the function the be an interface type that can hold the type context
when needed (see issue #1559)