move cutter and colbuidler out of proc
This commit moves the Cutter and ColumnBuilder out of the proc
package so they can be reused by expr and expr/function, which
will be needed by the cut() function.

ColumnBuilder was moved to its own package under zng/builder since
this purely a zng function to construct zng record values.  The Cutter
type has logic to take a stream of zng records and create "cuts" of
those value so this more complicated machinery was moved to package
expr.