zqd: Multipart log ingest (#1336)
Allow ingest of log data through streaming multi-part http request.

The files are submitted as an http multipart request. JSON typing
can be submitted as part of the request- it should be given the
form value of "json_type_config".

For http 1.1 the server cannot send responses back while the
requester is sending data. As a result remove status updates and
transmit any wires in the server responses. Clients that want to
know the amount of data sent while the request is in flight will
have to keep track the bytes written over the wire themselves.

The zapi post commands have been reformatted. They are as such:

- postpcap - post a pcap file to a space
- postpath - post paths that will be opened and read by zqd.
- post - stream log data to zqd over the network.

Closes #1084