remove uneeded match-all filter at front of flow graph
This commits changes the zql parser so it doesn't insert a filter proc
with a match-all op at the front of flowgraphs that do not have a filter.
We used to rely on this but that seems fragile and we only needed to
make a couple changes here to fix things.

One change was to add logic to the worker clustering requests so that
a filter with matchall is added.  A better fix would be to change the
search endpoint to handle a null Proc and create a scanner for the
given spans instead of panic.  The panic arose from a null Proc
causing a null pointer deref in ast.Unpack.  We fixed this with checks
that return an error instead of panic stacktrace.

This change is needed so that we can add a parallel input flag to zq,
feeding a parallel zql query as the inserted "filter *" gets in the way
and we need to know the difference between the user typing "* | ..."
vs "...".  This parallel input flag will be in turn needed by the
soon-coming merge join proc.