remove unneeded match-all filter at front of flow graph (#1618)
This commit changes the zql parser so it doesn't insert a filter proc
with a match-all op at the front of flow graphs that do not have a filter.
We used to rely on this but that seems fragile and we only needed to
make a couple changes here to fix things.

We also added error checks in the ast unpacker to detect null values 
inside of json nodes that would otherwise cause panics.

This change is needed so that we can add a parallel input flag to zq,
feeding a parallel zql query as the inserted "filter *" gets in the way
and we need to know the difference between the user typing "* | ..."
vs "...".  This parallel input flag will be in turn needed by the
soon-coming merge join proc.