add preliminary join proc (#1632)
This commit adds a join proc.  It works with the -P flag in zq
and supports just left joins.  Note that the semantics of a left
join are a bit different than sql in that when the expression for
the left key refers to a field that is not present, the record
is dropped from the join.  However, for right keys that are not
matched with the left key, the left record is included in the
result, as with expected left-join semantics.  There is no attempt
to put nulls in the columns for missing righthand keys (though
this could be added perhaps as an option to join).

It will be trivial to add right join and inner join support given
the current framework but I wanted to start with left join to
keep things simple at first while we gather experience and user
feedback.

The left and right inputs must be pre-sorted by the key expressions
or the results are undefined.  We may want to call this version of
join "mergejoin" as an explicit indicator to the query planner that
the operator knows the inputs will be sorted and its doesn't ahve to
do anything special.  Likewise, a vanilla version of join would be
smart enough to use information from the storage layer to determine
automatically which datapaths needed to be sorted and how.

Also, we'll want to add hash joins at some point, but merge joins
should get us pretty far, and we'll want to perform aggregations
while performing the join though this can be an under-the-hood
optimization where aggregations are pushed into the join operator.