zqd auto compaction
A simple first pass at providing auto-compaction functionality to zqd with
a simple hueristic: A compaction is run when an archivestore is first opened or
after data has been written to the archivestore. If new data is written to the
archive store while a compaction is running the current compaction task will be
canceled.

Closes #1555