add support for zql comments, multiline queryies, and zq -z
This commit adds support for zq -z to take a zql source file instead
of writing the query on the shell command line.  We also enhanced
the grammar so that queries may be split over multiple lines and
added support for "//"-style single-line comments.  Multi-line
comments a la "/* ... */" will come in a subsequent PR but we left
some placeholder grammar rules in anticipation of this.

Closes #1646