fix join key expressions and add test coverage
This commit fixes a bug in the join-key expression grammar
and adds a test to cover it.  There was ambiguity that was not
properly tested in the previous commit.  The ambiguity is resolved
by using parenthesis to specify an expression for the join key
instead of a direct field reference.  There may be a better
syntactic solution but it will need some creative thought and
we can live with this for now.

We also added support for simple form "join key", which doesn't
do anything but copy the left records that have the given key
expression, but will do a proper inner join when we add the
inner flag.