zson spec
This commit adds a markdown description of the new ZSON format.
References to tzng have been removed from the ZNG spec and replaced
with references to the ZSON spec.  The language in the ZNG spec
has been further simplified a bit assuming knowledge of the ZSON
data model described in the new spec.

The zsonio writer has been improved to fix bugs and follow a few
changes that happend to the ZSON format.

We made a first cut at updating the Zeek compat doc but we need to
do have the ZSON reader before this can be finished.  We should also
consider explaining compat in terms of ZSON instead of ZNG.