unify the three variants of AST (#1665)
This commit unifies the parsed ASTs so that whether an AST is
generated by pigeon, pegjs, or by marshaling an ast.Proc, the
resulting JSON is deep equal for the same zql query text.
Previously, different versions treated the zero and empty values
differently.  This involved changes to the peg grammar actions
as well as the go ast package.  We also updated the ast command
so that it can display these three variations, as previously,
there was no way to display the pigeon version of the ast.

We udpated the unit tests to compare the three variants and
added a test that looks through all the ztest yaml files to find
zqls to run and compare over the three AST variants.

Fixes #1638