Move archive Chunk to its own package (#1685)
Overtime package archive is becoming more complex with more moving parts.
Moving Chunk into its own package is an attempt at simplifying things. A chunk
is only aware of the directory it is supposed to write itself in. The wide
archive that is made up of chunks is responsible for writing them to the
correct tsdir.

Having a simplified chunk will allow for more isolated testing on specific
chunk features.