add experimental ZSON writer (#1681)
This commit adds a first cut at the ZSON writer.  It came together quite
nicely and is ready for people to try out and provide feedback.
In subsequent PRs, we'll add an update to the spec (issue #1678),
comprehensive tests, reader support (issue #1680), and misc other
finishes like enums and proper string escaping.