add network_of() function
This commit adds a function to return a network from an IP address.
If one argument is given, then its an IP address and that IP must be
an IPv4 address, which determines the default netmask to use via
the IP class A-B-C system.  If two arguments are given, then the
second argument is a netmask given either as a integer (indicating
the number of bits to use in the netmask, e.g., a /28 subnet)
or it's a network value in which case the network number must have
all network bits set to 1 and host bits set to 0, as in, 10.1.2.128/25.