simplify the groupby proc
This commit simplifies the groupby proc by using data structures from
expr and not making new aggregation evaluators for each row since the
evaluators depend only on the field names.

This is just code simplification and doesn't change any logic.

This will get groupby ready for the fix for issue #1701.