allow len() function to work on IPs (#1725)
This commit changes the len function to return the length of IP
addresses and networks.  This is handy for differentiating between
IPv4 (len=4) and IPv6 (len=16).  Likewise, networks are len 8 and
len 32.

We also fixed a bug in zng.EncodeNet where ipv6 networks were 
accidentally smashed to 0 for the IP part of the network value.