eliminate extra copy of groupby keys (#1732)
Previously, the groupby code stored the key values in the Go map
as its string key and stored a second copy of the key vlues in the
groupby.Row value of the Go map.  This commit gets rid of the second
copy and uses the Go map hash key as the source for the key values.