zson spec (#1715)
This commit adds a markdown description of the new ZSON format.
References to tzng have been removed from the ZNG spec and replaced
with references to the ZSON spec.  The language in the ZNG spec
has been further simplified a bit assuming knowledge of the ZSON
data model described in the new spec.

The zsonio writer has been improved to fix bugs and follow a few
changes that happend to the ZSON format.

We also updated the Zeek compat doc and reworded this as interoperability 
instead of compatibility.