eliminate zng.ErrUnset
Now that the zson spec more clearly defines the null concept,
the concept of zng.ErrUnset is no longer needed.

This commit removes zng.ErrUnset and has each zng.Decode*()
function return the zero value of the type it decodes to.
If a callee of zng.Decode*() wants to treat null differently
from the zero value, then it should check for zv.Bytes==nil
before calling zng.Decode*().  This means that ErrUnset is no
longer bubbled up as an error condition, e.g., to groupby,
and further, that we should decide what each functions should
do with null (the reducers already handle nulls).  For example,
len(null) returns int64(0) but replace("foo", null, null) is
an error.  I made a first pass over the expr/function in this
regard, but we should decide and document it all. See #1737.