Archive: use index package
Use the newly created index package to create archive indices. When a new index
rule is added, the rule is assigned a unique identifier then persistently
stored in the archive's indexdefs directory.

The pr aims to match feature parity of archive indices in their current form.
A future pr will introduce functionality to create new indices from index defs
on import of new data and rebuild indices when compactions are run.

Part of #1656