Fix microindex trailer read bug
The read trailer part of microindex was relying on the call to the
io.Reader.Read to fill the entire buffer. For the aws sdk this often do not
happen.

According to the golang io.Reader.Read documentation:

If some data is available but not len(p) bytes, Read conventionally returns
what is available instead of waiting for more.

Accordingly use io.ReadFull to ensure the entire trailer is read.

Closes #1730