finish refactoring compiler
This is the final, heavy lift on the compiler refactor issue #1162.
Except for the creation of the filter.Program interface and its
implementation as compiler.Filter, everything here is just code
movement.  All the compiler related logic from packages filter and
expr have been moved into package compiler.  Also, we missed
the rename and join procs in the previous commit so they are taken
care of here.

The filter.Program interface needed to be created to break an import
cycle since the scanner was previously calling filter.Compile and
moving that into compiler created the cycle.

Closes #1162.