fix bug in groupby output for dynamic types (#1740)
This commit fixes a bug where output types can change based on
input values so the assumption that input type ID determines the
output type record was invalid.  The fix is to map the vector
of types that comprise the output type to a unique small integer
and use the integer to index a map that provides the zng.TypeRecord
from the given type vector.  Since output types are typically one
maybe two distinct type values, the overhead of the linear-scan
lookup should be small.  Since key types can also change (this
was previously handled), we use the same type vector method
in the key values to simplify the code.

Fixes #1701.