intial zson parser and reader
This commit adds a parser and reader to package zson.  The spec
has been updated a bit reflecting lessons learned during the
implementation of the parser.

The type system in ZSON requires an interesting semantic analysis
of each ZSON value object.  The parser scans the input text and
parses it into an AST.  The AST is then traversed by the semantic
analysis pass to convert the AST into an intermediate value form.
This pass uses the distributed type information in the AST to
determine an exact type for each constituent value and typedef
in a nested value.  This fully-typed value is then passed to the
Builer and transformed into a zng represenation with a simple
traversal of the nested value.

The new ZSON ast nodes in ast/zson.go have been designed to be
zql-ready for the peg grammar to generate and pass to the search
API as zql queries with embedded ZSON constants of arbitrary
complexity and size.

A large suite of "boomerang" tests was created by traversing
the zq repo and finding any ztest yaml file that had tzng in it.
For each such hit, a file was created in zson/ztests with
generic boilerplate that converts the tzng to zson and the
zson is read as input and output again as zson making sure that
the two zson outputs are the same.