Fix tsDirWriter Flush race (#1766)
Fix a race occurring with tsDir Flush. If the stale flusher loop calls flush on
a tsdir at the same time the main write thread flushes that same tsdir, there's
a race ending with the loser getting for calling Spill on a nil
spill.MergeSort.

Fix this by placing a general mutex on both the Write() and flushStaleDirs()
calls of import.Writer. From my measurements this change has a negligible
change to import performance.

Closes #1759