improve zson types and add zson methods to zng.Type (#1765)
This commit improves the type system for zson.  It updates the spec
and improves the implementation.  We also added ZSON() and ZSONOf()
methods to zng.Type.  Support for enums is now implemented.

The TypeOfType implementation has been augmented with a hash table 
in package zson to convert between ZSON type-value strings and zng.Types.
This bypasses the need to parse a type string or format a string for a 
zng.Type for every type value encountered  (the parsing direction is not yet
implemented but is coming soon).

Closes #1675.