Recruiter API in zqd for server and client  (#1695)
This establishes the pattern of a zqd root process calling the recruiter to get a set of zqd workers, that were previously registered, so zqd root can use the set of worker for a distributed query. It has allowed us to do some preliminary scalability testing. This version is not usable when a K8s cluster is scaled up and down, unless the recruiter and workers are restarted. My next PR will add features to the recruiter to allow for cluster scaling without restarts.