zqd PostgresDB
In addition to storing SpaceRows using a filedb, add support for storing
SpaceRows with postgres. In package apiserver create PostgresDB struct that
implements the DB interface.

A lot of plumbing in this pr to make this work with respect to tests; ran both
locally and in ci:

- Add pgctl command for running common tasks on a zqd postgresdb. Currently this
has only options for creating a destroying a testdb, but eventually will be
extended to run migrations and introspect a database.
- zqd listen: Add -postgres arg for specifying database. Arg -postgres.DB also
added to specify database in instance (this will mostly be used for tests).
- Use package golang-migrate/migrate to create migrations folder in
ppl/zqd/migrations
- Add startup.sh script: To be used for postgres+zqd ztests. Uses pgctl to
create and destroy a testdb then launches zqd listen with the correct
configuration
- ztest: Plumbing for specifing environment variabes for shell scripts.
- ztest: Add tag feature. ztest yaml files that use the tag field will only run
if the tag matches a ZTEST_TAG env variable. This was needed so make test-system
didn't end up running postgres tests.

Closes #1752