more flexibe zson type names and aliased rec bug fix
This commit changes the definition of zson type names to include
the characters "." and "/".  This will permit go-style etc type
names of the form "github.com/acme/foo.Bar".  It will also be
useful to embed schema versions in type names for a new style of
decentralized and hierarchical schema management that zson
enables (e.g., shopping_cart_v1.0 or employee_record.2).

It also fixes a bug that happened when casting a union type
in alias form, which caused the alias to be lost and replaced
with the underlying union type.

Finally, we found and fixed a bug where r.field failed if r
was type alias of a record type.