create zson.Formatter from most of zsonio.Writer (#1840)
This commit creates a zson.Formatter from the logic that used to
live in zsonio.Writer.  With this movement, a coming-soon
zson.Marshaler can use the Formatter without creating an
import cycles.

This is mostly code movement and rename.  The one logic change
is the formatter now uses a stack of strings.Builder instead of
saving the writer pointer in the stack when re-entering.  Also,
instead of writing the formatted output to an io.Writer, the
output is built in memory using a strings.Bulder.