add zson marshaler and clean up zng marshaler (#1854)
This commit adds a zson marshaler, which is basically a wrapper
around a zng marshaler that allows one to do marshaling strictly
with ZSON text format and not worry about zng or package zng/resolver.

This work also exposed some problems with the zng marshaler regarding
interfaces so several bug fixes there included here as well as
an extension of the zng marshaler to handle empty interface for
non-struct types.  We also made things work when records have
recursively invested interface types (that are resolved with
marshaler bindings).