Don't pass in flowgraph warnings channel from outside (#1865)
Previously, the driver package allowed clients to pass in a flowgraph
warnings channel. This was used by clients that created a channel for
use by a zbuf.WarningReader, and then passed that channel to a
flowgraph.

This commit changes zbuf.WarningReader to take a callback instead. It is
a slightly simpler API and when used in conjunction with the driver package,
the pattern is cleaner: the driver creates its warning channel
internally, and neither WarningReader nor the client plumbing
things together needs to know about the channel.