Ingest shaper
This commit adds support for a running Z "shaper" script on incoming
logs. This facility is for tasks such as transforming record
structurs, casting leaf types, or policing incoming types against an
organization schema. (Z script does not quite yet have the support for
much of this, but that is imminent).